repackaging: * Changed name of package to auspixdggs (also lowercase following python naming conventions for packages) and moved the engine and callable modules to subdirectories of the package * Added author attribution and contact to Joe and GA * Updated scripts and enablement_tagging to reference the new package * Renamed some enablement_tagging scripts to add .py file extension
