+++ README.md
@@ -4,6 +4,16 @@ The AusPIX DGGS is an Australian Government implimentation of the rHealpix DGGS 
 Included in this repository is the rHealPIX "engine" and a series of scripts that use AusPIX in a variety of ways.
 
 Additionally there are several 'callable' modules that perform common tasks in DGGS space. These are in the "callable_modules" folder.
+## Installation 
+
+`python setup.py install`  
+or   
+`pip install .`   
+or for development  
+`python setup.py develop`  
+or   
+`pip install -e .`
+
 
 
 ![rhealpix](https://user-images.githubusercontent.com/23160509/53066271-23aa4680-3523-11e9-8e6c-2f042f9befbf.png)
@@ -22,6 +32,8 @@ Source references for rHealPIX:
 Gibb, R.G., 2016, “The rHealPIX Discrete Global Grid System” Proceedings of the 9th Symposium of the International Society for Digital Earth (ISDE), Halifax, Nova Scotia, Canada. IOP Conf. Series: Earth and Environmental Science, 34, 012012. DOI: 10.1088/1755-1315/34/1/012012
 
 Gibb R G, Raichev A and Speth M 2013 The rHEALPix Discrete Global Grid System URL http://dx.doi.org/10.7931/J2D21VHM
+
+
