+++ enablement_tagging/firePolygons_example.py
@@ -6,8 +6,8 @@ Joseph Bell Geoscience Australia Jan 2020
 '''
 
 import shapefile  #to read and write shapefiles
-import dggs_in_poly  # this is script written by GA - see GitHub AusPix DGGS enablement folder
-from dggs import RHEALPixDGGS
+import auspixdggs.callablemodules.dggs_in_poly  # this is script written by GA - see GitHub AusPix DGGS enablement folder
+from auspixdggs.auspixengine.dggs import RHEALPixDGGS
 rdggs = RHEALPixDGGS() # make an instance
 
 # identify path to a shape file to get your polygons from
