+++ scripts/add_DGGS_Geonames.py
@@ -20,10 +20,10 @@ output = list()
 
 
 import csv
-from dggs import RHEALPixDGGS
+from auspixdggs.auspixengine.dggs import RHEALPixDGGS
 
 import math
-from utils import my_round
+from auspixdggs.auspixengine.utils import my_round
 
 # make an instance        
 rdggs = RHEALPixDGGS()
