+++ scripts/add_DGGS_RiverGs.py
@@ -22,10 +22,10 @@ output = list()
 
 
 import csv
-from dggs import RHEALPixDGGS
+from auspixdggs.auspixengine.dggs import RHEALPixDGGS
 
 import math
-from utils import my_round
+from auspixdggs.auspixengine.utils import my_round
 
 # make an instance        
 rdggs = RHEALPixDGGS()
